Implement custatevec simulation state subtype
