+++ runtime/nvqir/cutensornet/tensornet_state.h
@@ -12,6 +12,9 @@
 #include "timing_utils.h"
 #include <unordered_map>
 
+#include "common/EigenDense.h"
+#include "common/SimulationState.h"
+
 namespace nvqir {
 /// @brief Wrapper of cutensornetState_t to provide convenient API's for CUDAQ
 /// simulator implementation.
@@ -76,5 +79,126 @@ public:
 
   /// @brief Destructor
   ~TensorNetState();
+};
+
+class TensorNetSimulationState : public cudaq::SimulationState {
+protected:
+  /// @brief Reference to the `TensorNetState` pointer. We
+  /// take ownership of it here.
+  TensorNetState *state;
+
+  /// @brief FIXME We should remove this. For not it is for backward
+  /// compatibility
+  const std::vector<cudaq::complex128> stateData;
+
+public:
+  TensorNetSimulationState(TensorNetState *inState)
+      : state(inState), stateData(inState->getStateVector()) {}
+
+  std::size_t getNumQubits() const override { return state->getNumQubits(); }
+
+  /// @brief Return the shape of the data.
+  std::vector<std::size_t> getDataShape() const override {
+    // FIXME I imagine this should be different
+    return {getNumQubits()};
+  }
+
+  /// @brief Compute the overlap of this state with the provided one.
+  /// If the other state is not on GPU device, this function will
+  /// copy the data from host.
+  double overlap(const cudaq::SimulationState &other) override {
+    if (other.getDataShape() != getDataShape())
+      throw std::runtime_error("[tensornet-state] overlap error - other state "
+                               "dimension not equal to this state dimension.");
+
+    if (other.isDeviceData())
+      throw std::runtime_error("[tensornet-state] cannot compute "
+                               "overlap with GPU state data yet.");
+
+    return std::abs(
+        Eigen::Map<Eigen::VectorXcd>(
+            const_cast<cudaq::complex128 *>(stateData.data()), stateData.size())
+            .transpose()
+            .dot(Eigen::Map<Eigen::VectorXcd>(
+                reinterpret_cast<cudaq::complex128 *>(other.ptr()),
+                stateData.size()))
+            .real());
+  }
+
+  /// @brief Compute the overlap of this state with the data provided as a
+  /// `std::vector<double>`. If this device state is not FP64, throw an
+  /// exception. This overload requires an explicit copy from host memory.
+  double overlap(const std::vector<cudaq::complex128> &other) override {
+    if (stateData.size() != other.size())
+      throw std::runtime_error("[tensornet-state] overlap error - other state "
+                               "dimension not equal to this state dimension.");
+    return std::abs(
+        Eigen::Map<Eigen::VectorXcd>(
+            const_cast<cudaq::complex128 *>(stateData.data()), stateData.size())
+            .transpose()
+            .dot(Eigen::Map<Eigen::VectorXcd>(
+                reinterpret_cast<cudaq::complex128 *>(
+                    const_cast<cudaq::complex128 *>(other.data())),
+                other.size()))
+            .real());
+  }
+
+  /// @brief Compute the overlap of this state with the data provided as a
+  /// `std::vector<float>`. If this device state is not FP32, throw an
+  /// exception. This overload requires an explicit copy from host memory.
+  double overlap(const std::vector<cudaq::complex64> &other) override {
+    throw std::runtime_error(
+        "[tensornet-state] requires FP64 data for overlap computation.");
+  }
+
+  /// @brief Compute the overlap of this state with the data provided as a raw
+  /// pointer. This overload will check if this pointer corresponds to a device
+  /// pointer. It will copy the data from host to device if necessary.
+  double overlap(cudaq::complex128 *other, std::size_t numElements) override {
+    if (stateData.size() != numElements)
+      throw std::runtime_error("[tensornet-state] overlap error - other state "
+                               "dimension not equal to this state dimension.");
+
+    return std::abs(
+        Eigen::Map<Eigen::VectorXcd>(
+            const_cast<cudaq::complex128 *>(stateData.data()), stateData.size())
+            .transpose()
+            .dot(Eigen::Map<Eigen::VectorXcd>(other, numElements))
+            .real());
+  }
+
+  double overlap(cudaq::complex64 *other, std::size_t numElements) override {
+    throw std::runtime_error(
+        "[tensornet-state] requires FP64 data for overlap computation.");
+  }
+
+  /// @brief Return the vector element at the given index.
+  cudaq::complex128 vectorElement(std::size_t idx) override {
+    return stateData[idx];
+  }
+
+  /// @brief Dump the state to the given output stream
+  void dump(std::ostream &os) const override {
+    os << Eigen::Map<Eigen::VectorXcd>(
+              const_cast<cudaq::complex128 *>(stateData.data()),
+              stateData.size())
+       << "\n";
+  }
+
+  /// @brief This state is GPU device data, always return true.
+  bool isDeviceData() const override { return false; }
+
+  /// @brief Return the raw pointer to the device data.
+  void *ptr() const override {
+    return const_cast<cudaq::complex128 *>(stateData.data());
+  }
+
+  /// @brief Return the precision of the state data elements.
+  precision getPrecision() const override {
+    return cudaq::SimulationState::precision::fp64;
+  }
+
+  /// @brief Free the device data.
+  void destroyState() override { delete state; }
 };
 } // namespace nvqir
